requirements.txt
python-telegram-bot==21.4
boto3
nest_asyncio
